+++ Assets/Packages/DeviceDebugger/CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
 
+
+## [1.0.1] - 2025-09-25
+- Remove [MenuItem("Tools/Graphy/Create Prefab Variant")]
+- Remove [MenuItem( "Tools/Graphy/Import Graphy Customization Scene" )]
+- Add [MenuItem("GameObject/Device Debugger")]
+- Add [MenuItem("Tools/Device Debugger/Add to Scene")]
+
 ## [0.0.1] - 2025-08-15
 - Create new